requirements.txt
requests
lxml
pyyaml
python-telegram-bot
dingtalkchatbot
googletrans==4.0.0rc1
